# Improve block querying with a binary join and sync iterator