@simonswine 2023-07-10
The difference in the `ColumnIterator` vs `SyncIterator` is minuscule:

```
❯ benchstat before.txt after.txt
name                    old time/op    new time/op    delta
ColumnIterator/sync-12    17.9ms ± 5%    17.4ms ± 4%  -3.14%  (p=0.007 n=10+10)

name                    old alloc/op   new alloc/op   delta
ColumnIterator/sync-12    68.6MB ± 0%    65.7MB ± 0%  -4.25%  (p=0.000 n=10+10)

name                    old allocs/op  new allocs/op  delta
ColumnIterator/sync-12      200k ± 0%      200k ± 0%  -0.09%  (p=0.000 n=10+9)
```

But getting rid of the async aspect is beneficial stopping overreading.

A bigger impact is on the `JoinIterator` vs `BinaryJoinIterator` and the amount of pages read from a block:
- It will now seekOrNext as soon as possible rather than reading the first page of every column and the move on.
- Also it is quite beneficial doing a next instead of sync if the it is the adjacent row (that might be something we would rather have in the SyncIterator maybe)

Left --- / Right +++

![image](https://github.com/grafana/phlare/assets/223048/09439eb6-9eab-4204-90f1-111612aa146c)


The impact to the overall query time is no longer so significant, this was probably due a bug that missed out on some matches.